CameraIdle added into game
If player doesn't control the camera for a handful of seconds, it will reset to players current direction, and smoothly update and follow player.

Current bumps to eventually fix: Looking backwards, had to ignore code when looking directly backwards, or it would violently jitter back and forth. So it has a deadspot at our 6, and looking out of it will jolt the camera quickly, So i'll need to manually slow it down around 170 deg and up. But can't put more work into it atm, so getting it up and ready for feedback

Also made in a way that can be adapted to objective goal viewing, with a lot of tweaking, but doable in near future.

Only files changed are the character cpp and h files